requirements.txt
asgiref==3.4.1
astroid==2.8.6
black==21.11b1
click==8.0.3
coverage==6.1.2
Django==3.2.8
djangorestframework==3.12.4
flake8==4.0.1
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.11.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
regex==2021.11.10
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
wrapt==1.13.3
